requirements.txt
numpy==1.20.3
pandas==1.3.5
Pillow==9.0.1
plotly==5.5.0
scikit_learn==1.0.2
streamlit==1.4.0
click<=8.0.4
